Litle change

+++ Client/www/js/controllers.js
@@ -14,14 +14,17 @@ angular.module('App.controllers', [])
 		});
 		
 		$scope.$watchCollection('products', function() {
-			$scope.nbProducts = Object.keys($scope.products).length;
+			if ($scope.products != undefined) $scope.nbProducts = Object.keys($scope.products).length;
+			else $scope.nbProducts = 0;
 		});
 		
         $scope.addProduct = function() {
             //TODO check if mobile to enter the function ! Otherwise it throws error about cordova undefined
             cordova.plugins.barcodeScanner.scan(function (result) {
-                    if ($scope.products[result.text] != undefined) $scope.products[result.text].plus();
-                    else Datas.addProductToFridge(new Product(result.text));
+					if (result.text.length > 0) {
+						if ($scope.products[result.text] != undefined) $scope.products[result.text].plus();
+						else Datas.addProductToFridge(new Product(result.text));
+					}
                 }, function (error) {
                     alert("Scanning failed: " + error);
                 }
@@ -48,20 +51,21 @@ angular.module('App.controllers', [])
         }
     }])
 
-    .controller('MarketCtrl', ['$scope', 'Datas', 'Product', function ($scope, Datas, Product) {
+    .controller('MarketCtrl', ['$scope', 'Datas', 'Product', 'DialogProduct', function ($scope, Datas, Product, DialogProduct) {
         $scope.products = Datas.getMarket().then(function(data) {
 			$scope.products = data;
 		});
 		
 		$scope.$watchCollection('products', function() {
-			$scope.nbProducts = Object.keys($scope.products).length;
+			if ($scope.products != undefined) $scope.nbProducts = Object.keys($scope.products).length;
+			else $scope.nbProducts = 0;
 		});
 
         $scope.addProduct = function() {
             //TODO check if mobile to enter the function ! Otherwise it throws error about cordova undefined
             cordova.plugins.barcodeScanner.scan(function (result) {
                     if ($scope.products[result.text] != undefined) $scope.products[result.text].plus();
-                    else Datas.addProductToMarket(new Product(result.text));
+                    else Datas.addProductToMarket(new Product(result.text, 1, true));
                 }, function (error) {
                     alert("Scanning failed: " + error);
                 }
@@ -109,11 +113,13 @@ angular.module('App.controllers', [])
         }
 
         $scope.delete = function() {
-            $scope.product.deleteFromDb();
-            if ($scope.product.isMarket()) Datas.removeProductFromMarket($scope.product);
-            else Datas.removeProductFromFridge($scope.product);
+			if (confirm("Remove the product ?")) {
+				$scope.product.deleteFromDb();
+				if ($scope.product.isMarket()) Datas.removeProductFromMarket($scope.product);
+				else Datas.removeProductFromFridge($scope.product);
 
-            $mdDialog.hide();
+				$mdDialog.hide();
+			}
         }
 
         $scope.close = function() {
